+++ spec/osm/evening_spec.rb
@@ -118,6 +118,35 @@ describe "Evening" do
     end
 
 
+    it "Add activity to evening (succeded)" do
+      url = 'https://www.onlinescoutmanager.co.uk/programme.php?action=addActivityToProgramme'
+      post_data = {
+        'apiid' => @CONFIGURATION[:api][:osm][:id],
+        'token' => @CONFIGURATION[:api][:osm][:token],
+        'userid' => 'user_id',
+        'secret' => 'secret',
+        'meetingdate' => '2000-01-02',
+        'sectionid' => 1,
+        'activityid' => 2,
+        'notes' => 'Notes',
+      }
+
+      Osm::Term.stub(:get_for_section) { [] }
+      HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"result":0}'}) }
+      activity = Osm::Activity.new(:id => 2, :title => 'Title')
+      evening = Osm::Evening.new(:section_id => 1, :meeting_date => Date.new(2000, 1, 2))
+      evening.add_activity(@api, activity, 'Notes').should be_true
+      evening.activities[0].activity_id.should == 2
+    end
+
+    it "Add activity to evening (failed)" do
+      HTTParty.should_receive(:post) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"result":1}'}) }
+      activity = Osm::Activity.new(:id => 2, :title => 'Title')
+      evening = Osm::Evening.new(:section_id => 1, :meeting_date => Date.new(2000, 1, 2))
+      evening.add_activity(@api, activity, 'Notes').should be_false
+    end
+
+
     it "Update an evening (succeded)" do
       url = 'https://www.onlinescoutmanager.co.uk/programme.php?action=editEvening'
       post_data = {